add the script

+++ generalframework/utils/utils.py
@@ -247,7 +247,7 @@ def save_images(segs: Tensor, names: Iterable[str], root: Union[str, Path], mode
 
             save_path.parent.mkdir(parents=True, exist_ok=True)
 
-            imsave(str(save_path), seg.cpu().numpy())
+            imsave(str(save_path), seg.cpu().numpy().astype(np.uint8))
 
 
 # dataset
